Changed postImageController.js to be able to handle requests sent from Dragger component Changed imageUpload component so that the methods it uses do not need to be passed through Changed imageUpload component to return a class instead of a function
